fix: lint errors
