+++ src/extension/options-page/SetupStep.ts
@@ -1,0 +1,8 @@
+export enum SetupStep {
+    ConsentDataCollection = 'consent-data-collection',
+    CreatePersona = 'create-persona',
+    ConnectNetwork = 'connect-network',
+    RestoreDatabase = 'restore-database',
+    RestoreDatabaseAdvance = 'restore-database-advance',
+    RestoreDatabaseConfirmation = 'restore-database-confirmation',
+}
