+++ src/extension/options-page/index.tsx
@@ -25,7 +25,7 @@ import DashboardWalletsRouter from './DashboardRouters/Wallets'
 import DashboardContactsRouter from './DashboardRouters/Contacts'
 import DashboardSettingsRouter from './DashboardRouters/Settings'
 import { appearanceSettings, Appearance } from '../../components/shared-settings/settings'
-import { DashboardSetupRouter, SetupStep } from './DashboardRouters/Setup'
+import { DashboardSetupRouter } from './DashboardRouters/Setup'
 import { DashboardBlurContextUI } from './DashboardContexts/BlurContext'
 import { DashboardRoute } from './Route'
 import { SSRRenderer } from '../../utils/SSRRenderer'
@@ -38,6 +38,7 @@ import Services from '../service'
 import { RequestPermissionPage } from '../../components/RequestPermission/RequestPermission'
 import { grey } from '@material-ui/core/colors'
 import { DashboardSnackbarProvider } from './DashboardComponents/DashboardSnackbar'
+import { SetupStep } from './SetupStep'
 
 const useStyles = makeStyles((theme) => {
     const dark = theme.palette.type === 'dark'
